+++ common/src/main/java/org/geogebra/common/kernel/prover/AbstractProverReciosMethod.java
@@ -225,11 +225,11 @@ public abstract class AbstractProverReciosMethod {
 			SymbolicParameters s, AlgebraicStatement as) {
 		if (as != null) {
 			// use Botana's method
-			HashMap<PVariable, Long> substitutions = new HashMap<PVariable, Long>();
+			HashMap<PVariable, BigInteger> substitutions = new HashMap<PVariable, BigInteger>();
 			for (Entry<PVariable, BigInteger> entry : values.entrySet()) {
 				PVariable v = entry.getKey();
 				// FIXME: Change Long in Variable to BigInteger
-				substitutions.put(v, entry.getValue().longValue());
+				substitutions.put(v, entry.getValue());
 			}
 			ProverSettings proverSettings = ProverSettings.get();
 			ExtendedBoolean solvable = PPolynomial.solvable(
@@ -264,11 +264,11 @@ public abstract class AbstractProverReciosMethod {
 			values.put(variable, BigInteger.valueOf(i));
 			if (as != null) {
 				// use Botana's method
-				HashMap<PVariable, Long> substitutions = new HashMap<PVariable, Long>();
+				HashMap<PVariable, BigInteger> substitutions = new HashMap<PVariable, BigInteger>();
 				for (Entry<PVariable, BigInteger> entry : values.entrySet()) {
 					PVariable v = entry.getKey();
 					// FIXME: Change Long in Variable to BigInteger
-					substitutions.put(v, entry.getValue().longValue());
+					substitutions.put(v, entry.getValue());
 				}
 				ProverSettings proverSettings = ProverSettings.get();
 				ExtendedBoolean solvable = PPolynomial.solvable(
@@ -318,12 +318,12 @@ public abstract class AbstractProverReciosMethod {
 
 				if (as != null) {
 					// use Botana's method
-					HashMap<PVariable, Long> substitutions = new HashMap<PVariable, Long>();
+					HashMap<PVariable, BigInteger> substitutions = new HashMap<PVariable, BigInteger>();
 					for (Entry<PVariable, BigInteger> entry : values
 							.entrySet()) {
 						PVariable v = entry.getKey();
 						// FIXME: Change Long in Variable to BigInteger
-						substitutions.put(v, entry.getValue().longValue());
+						substitutions.put(v, entry.getValue());
 					}
 					ExtendedBoolean solvable = PPolynomial.solvable(
 							as.getPolynomials()
