+++ desktop/src/main/java/org/geogebra/desktop/kernel/prover/ProverReciosMethodD.java
@@ -220,12 +220,12 @@ public class ProverReciosMethodD extends AbstractProverReciosMethod {
 
 			if (as != null) {
 				// use Botana's method
-				HashMap<PVariable, Long> substitutions = new HashMap<PVariable, Long>();
+				HashMap<PVariable, BigInteger> substitutions = new HashMap<PVariable, BigInteger>();
 				for (Entry<PVariable, BigInteger> entry : values.entrySet()) {
 
 					PVariable v = entry.getKey();
 					// FIXME: Change Long in Variable to BigInteger
-					substitutions.put(v, entry.getValue().longValue());
+					substitutions.put(v, entry.getValue());
 				}
 				ExtendedBoolean solvable = PPolynomial.solvable(
 						as.getPolynomials()
